test: refactor `Datastore\instance` metric tests
Use integration_runner's ability to expand environmental variables in
expected data to ensure that tests with the `Datastore/instance/` in
EXPECT_METRIC pass no matter where the data store is hosted: localhost
or remote server (eg. a service in a container).

needs-work: this needs to be applied to other redis and predis tests