ci: allow multiple symbol defs at agent test link
LDFLAGS cannot be set with make command argument because it overrides
its value defined in Makefile. This causes `-pthread` to be removed from
LDFLAGS and as a result test binary cannot be built because pthread
symbols are unresolved (no pthread library provided in the link command).
A new variable USER_LDFLAGS is introduced as a workaround for this issue.