build: fix pcre build dependency
Don't depend on pcre-config return information how to include libpcre in
the build because this always results in linking to a shared object which
becomes are a runtime dependency that is hard to satisfy because
different Linux distributions use different name of the shared object
(libpcre.so or libpcre3.so). To hedge against this nuisance, agent needs
to statically link to pcre library.  Therefore a new configuration
option is added to agent's configure script: --with-pcre. It allows to
provide path prefix to a lib directory that contains libpcre.a.  The
default location is /opt/nr/pcre/8.40 which is where build scripts
install this library.