build: relax dependency on protobuf-compiler
protobuf-compiler is needed to compile 8T protobufs for daemon. Give the
use an option to use user or system provided protobuf-compiler rather
than always building vendored subdir.

Please note that protobuf-compiler is not a critical dependency because
daemon-protobuf target needs manual installation of go package even if
protobuf-compiler is installed.