fix cat integration tests
CAT integration tests expect a valid X-Newrelic-App-Data to be returned
in the response headers. In order for this to happen, APP_supportability
needs to be listed as trusted account id. Inject APP_supportability
into trusted_account_ids in mock connect reply returned by mock daemon
to the agent (in a very dirty, no golang way - to be fixed).