tests(agent): fix test logic for 8.3 exceptions
PHP 8.3 will invoke the user_exception_handler when an exception is
encountered (if one is registered). This results in an inversion of the
expectation from pre-8.3  of failing out of the zend_try/zend_catch
block of code that sets the zend_result to FAILURE and returns NULL to
the expr variable assigned to nr_php_call().
8.3 logic allows our newrelic_exception_handler to intercept the
exception, skip the zend_catch, and return SUCCESS in
nr_php_call_user_func() which returns a non-NULL value.